Pick up node prune from the GOPATH

+++ pack.ps1
@@ -16,22 +16,28 @@ $buildDirectoryPath = "$PSScriptRoot/dist"
 $buildArtifactsPath = "$buildDirectoryPath/Artifacts"
 
 function CleanNodeModules() {
+    $command = "node-prune.exe";
+
     if ((Get-Command node-prune -ErrorAction SilentlyContinue) -eq $null)
     {
-        Write-Error "Install go and then install node-prune (https://github.com/tj/node-prune)"
-        Write-Error "go get github.com/tj/node-prune/cmd/node-prune"
-        Exit 1
-    }
-    node-prune $PSScriptRoot\node_modules
-    node-prune $PSScriptRoot\source\node_modules
-    node-prune $PSScriptRoot\dist\tasks\CreateOctopusRelease\node_modules
-    node-prune $PSScriptRoot\dist\tasks\Deploy\node_modules
-    node-prune $PSScriptRoot\dist\tasks\OctoCli\node_modules
-    node-prune $PSScriptRoot\dist\tasks\OctoInstaller\node_modules
-    node-prune $PSScriptRoot\dist\tasks\Pack\node_modules
-    node-prune $PSScriptRoot\dist\tasks\Promote\node_modules
-    node-prune $PSScriptRoot\dist\tasks\Push\node_modules
+        $command = "$($env:GOPATH)\bin\node-prune.exe"
 
+        if(-Not (Test-Path $command)){
+            Write-Error "Install go and then install node-prune (https://github.com/tj/node-prune)"
+            Write-Error "go get github.com/tj/node-prune/cmd/node-prune"
+            Exit 1
+        }
+    }
+
+    Invoke-Expression "$command $($PSScriptRoot)\node_modules"
+    Invoke-Expression "$command $($PSScriptRoot)\source\node_modules"
+    Invoke-Expression "$command $($PSScriptRoot)\dist\tasks\CreateOctopusRelease\node_modules"
+    Invoke-Expression "$command $($PSScriptRoot)\dist\tasks\Deploy\node_modules"
+    Invoke-Expression "$command $($PSScriptRoot)\dist\tasks\OctoCli\node_modules"
+    Invoke-Expression "$command $($PSScriptRoot)\dist\tasks\OctoInstaller\node_modules"
+    Invoke-Expression "$command $($PSScriptRoot)\dist\tasks\Pack\node_modules"
+    Invoke-Expression "$command $($PSScriptRoot)\dist\tasks\Promote\node_modules"
+    Invoke-Expression "$command $($PSScriptRoot)\dist\tasks\Push\node_modules"
 }
 
 function UpdateTfxCli() {
